pipeline/requirements.txt
psycopg2-binary>=2.9.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dateutil>=2.8.2
unicodedata2>=15.0.0
